refactor(agent-sdk): remove dead public surface; neutral home for shared resolver helpers
PromptLoader, ModelResolver/ModelResolutionException, and
ResourceLocalCache/LocalCacheException had zero callers or
implementations anywhere in the repo. The descriptor/required/string
validation helpers move from HttpToolResolver to ToolRefAttributes so
the Java and MCP resolvers stop depending on the HTTP resolver class.
The DeepAgent placeholder comment no longer claims a dependency version
that contradicts the pom.
