--- agent-sdk/src/main/java/com/huawei/ascend/agentsdk/spec/model/ModelResolutionException.java
@@ -1,10 +1,0 @@
-package com.huawei.ascend.agentsdk.spec.model;
-
-import com.huawei.ascend.agentsdk.support.AgentSdkException;
-
-public class ModelResolutionException extends AgentSdkException {
-    public ModelResolutionException(String message) {
-        super(message);
-    }
-}
-
