--- agent-sdk/src/main/java/com/huawei/ascend/agentsdk/spec/model/ModelResolver.java
@@ -1,6 +1,0 @@
-package com.huawei.ascend.agentsdk.spec.model;
-
-public interface ModelResolver {
-    Object resolve(ModelSpec modelSpec);
-}
-
