--- agent-sdk/src/main/java/com/huawei/ascend/agentsdk/spec/prompt/PromptLoader.java
@@ -1,8 +1,0 @@
-package com.huawei.ascend.agentsdk.spec.prompt;
-
-public final class PromptLoader {
-    public PromptSpec inlineSystem(String system) {
-        return new PromptSpec(system == null ? "" : system);
-    }
-}
-
