--- agent-sdk/src/main/java/com/huawei/ascend/agentsdk/support/cache/LocalCacheException.java
@@ -1,10 +1,0 @@
-package com.huawei.ascend.agentsdk.support.cache;
-
-import com.huawei.ascend.agentsdk.support.AgentSdkException;
-
-public class LocalCacheException extends AgentSdkException {
-    public LocalCacheException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
